% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_coreact.R
\name{cbind.coreact_data}
\alias{cbind.coreact_data}
\title{Combine Coreact Data Objects (Column-bind)}
\usage{
\method{cbind}{coreact_data}(..., deparse.level = 1)
}
\arguments{
\item{...}{`coreact_data` objects to combine.}

\item{deparse.level}{Integer. Ignored.}
}
\value{
A combined `coreact_data` object.
}
\description{
Combines multiple coreact_data objects by adding samples.
}
