% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{coreact_pipeline}
\alias{coreact_pipeline}
\title{Run Full Coreact Pipeline (Disc-to-Disc)}
\usage{
coreact_pipeline(
  path_x,
  path_y,
  out_path,
  name_x = "Input_X",
  name_y = "Input_Y",
  meta_cols_x = 1,
  meta_cols_y = 1,
  min_prevalence = 0,
  filter_config = list(min_intersection = 1),
  fdr_threshold = 0.05,
  n_cores = 1,
  chunk_size = NULL
)
}
\arguments{
\item{path_x}{Path to input TSV X.}

\item{path_y}{Path to input TSV Y.}

\item{out_path}{Path for result file.}

\item{name_x}{Name for dataset X (default "Input_X").}

\item{name_y}{Name for dataset Y (default "Input_Y").}

\item{meta_cols_x}{Vector. Metadata columns for X.}

\item{meta_cols_y}{Vector. Metadata columns for Y.}

\item{min_prevalence}{Numeric vector (length 1 or 2).}

\item{filter_config}{List of structural filters.}

\item{fdr_threshold}{Numeric. Max FDR to retain (default 0.05).}

\item{n_cores}{Integer. Number of cores.}

\item{chunk_size}{Integer or NULL. Auto-calculated if NULL.}
}
\description{
Run Full Coreact Pipeline (Disc-to-Disc)
}
