% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_by_prevalence}
\alias{filter_by_prevalence}
\title{Filter Coreact Data by Prevalence (Internal)}
\usage{
filter_by_prevalence(obj, thresh)
}
\arguments{
\item{obj}{A `coreact_data` object.}

\item{thresh}{Numeric. The filter threshold.}
}
\value{
A subsetted `coreact_data` object.
}
\description{
Filters a coreact_data object based on row prevalence.
Automatically detects if the threshold is a percentage (< 1) or absolute count (>= 1).
}
\keyword{internal}
