% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_coreact.R
\name{filter_prevalence}
\alias{filter_prevalence}
\title{Filter features by prevalence}
\usage{
filter_prevalence(data, threshold)
}
\arguments{
\item{data}{A 'coreact_data' object.}

\item{threshold}{Numeric. Values less than 1 are interpreted as a fraction (percentage). Values greater than or equal to 1 are interpreted as absolute counts.}
}
\value{
A filtered 'coreact_data' object, or NULL if all features are removed.
}
\description{
Removes features that do not meet the specified prevalence threshold.
}
