% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_col_mismatch}
\alias{format_col_mismatch}
\title{Format Column Mismatch Error (Internal Helper)}
\usage{
format_col_mismatch(mat_x, mat_y, name_x, name_y)
}
\arguments{
\item{mat_x}{Matrix X}

\item{mat_y}{Matrix Y}

\item{name_x}{Name of dataset X}

\item{name_y}{Name of dataset Y}
}
\value{
A formatted error string detailing specific mismatches.
}
\description{
Format Column Mismatch Error (Internal Helper)
}
\keyword{internal}
