% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import_coreact_tsv}
\alias{import_coreact_tsv}
\title{Import TSV as Coreact Data Object}
\usage{
import_coreact_tsv(
  path,
  meta_cols = 1,
  feature_id = NULL,
  feature_id_sep = "|",
  name = "unknown",
  n_cores = 1,
  binarize = FALSE
)
}
\arguments{
\item{path}{String. Path to the .tsv file.}

\item{meta_cols}{Vector (integer or character). Indices or names of columns
containing metadata.}

\item{feature_id}{Vector (integer or character). The columns to use to construct
unique row names (Feature IDs). Must be a subset of `meta_cols`.
If multiple columns are provided, they are concatenated using `feature_id_sep`.
If NULL, row numbers are used.}

\item{feature_id_sep}{String. Separator used when concatenating multiple
feature_id columns. Default is "|".}

\item{name}{String. Identifier for the dataset.}

\item{n_cores}{Integer. Number of threads for fast reading.}

\item{binarize}{Logical. Should the data matrix be converted to binary?}
}
\value{
A 'coreact_data' object.
}
\description{
Reads a TSV, separates metadata columns from sample data, and converts
the sample data to a sparse matrix.
}
