% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_coreact.R
\name{print.coreact_data}
\alias{print.coreact_data}
\title{Print Method for Coreact Data}
\usage{
\method{print}{coreact_data}(x, ...)
}
\arguments{
\item{x}{A `coreact_data` object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Displays a summary of the Coreact object.
}
