% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_xy_args}
\alias{resolve_xy_args}
\title{Resolve X/Y Pipeline Arguments (Internal)}
\usage{
resolve_xy_args(val, arg_name)
}
\arguments{
\item{val}{The input argument (list or vector).}

\item{arg_name}{String. Name of the argument for error reporting.}
}
\value{
A list with elements `$x` and `$y`.
}
\description{
Handles the logic for arguments that can be provided as a single vector (applied to both)
or a list of length 2 (specific to X and Y).
}
\keyword{internal}
