% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{run_coreact}
\alias{run_coreact}
\title{Run Coreact Engine (Internal)}
\usage{
run_coreact(
  data_x,
  data_y,
  n_cores = 1,
  chunk_size = NULL,
  filter_config = list(min_intersection = 1),
  fdr_threshold = NULL
)
}
\description{
Run Coreact Engine (Internal)
}
