% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_coreact.R
\name{[.coreact_data}
\alias{[.coreact_data}
\title{Subset Coreact Data}
\usage{
\method{[}{coreact_data}(x, i, j, ...)
}
\arguments{
\item{x}{A `coreact_data` object.}

\item{i}{Row indices (features).}

\item{j}{Column indices (samples).}

\item{...}{Additional arguments (ignored).}
}
\description{
Subsetting operator.
}
