% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import_coreact_tsv}
\alias{import_coreact_tsv}
\title{Import TSV as Coreact Data Object}
\usage{
import_coreact_tsv(path, meta_cols = 1, name = "unknown", n_cores = 1)
}
\arguments{
\item{path}{String. Path to the .tsv file.}

\item{meta_cols}{Vector (names or indices) of metadata columns.}

\item{name}{String. Identifier for the dataset.}

\item{n_cores}{Integer. Number of threads for fast reading.}
}
\description{
Import TSV as Coreact Data Object
}
