% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{run_coreact}
\alias{run_coreact}
\title{Run Coreact Engine (Internal)}
\usage{
run_coreact(
  data_x,
  data_y,
  n_cores = 1,
  chunk_size = NULL,
  filter_config = list(min_intersection = 1)
)
}
\arguments{
\item{data_x, data_y}{coreact_data objects.}

\item{n_cores}{Integer.}

\item{chunk_size}{Integer.}

\item{filter_config}{List of thresholds.}
}
\value{
A tibble of results.
}
\description{
Run Coreact Engine (Internal)
}
