% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{worker_chunk_calc}
\alias{worker_chunk_calc}
\title{Internal Worker: Compute Co-occurrence Stats for a Chunk}
\usage{
worker_chunk_calc(
  idx_x_chunk,
  mat_x_chunk,
  mat_y,
  prev_x_chunk,
  prev_y,
  n_samples,
  config
)
}
\arguments{
\item{idx_x_chunk}{Vector of absolute indices for the rows of X in this chunk.}

\item{mat_x_chunk}{The actual data sub-matrix for X (dense or sparse).}

\item{mat_y}{The full matrix Y (dense or sparse).}

\item{prev_x_chunk}{Vector of prevalence (rowSums) for rows in X chunk.}

\item{prev_y}{Vector of prevalence (rowSums) for all rows in Y.}

\item{n_samples}{Integer, total number of samples (ncol).}

\item{config}{List of name-value pairs of filter thresholds. Current valid options are: `min_intersection`, `min_jaccard`, `min_overlap`.}
}
\value{
A tibble of passing pairs with stats.
}
\description{
Internal Worker: Compute Co-occurrence Stats for a Chunk
}
\keyword{internal}
